fix: make custom dialogue components their own files
