+++ client/components/Welcome.js
@@ -1,67 +1,16 @@
 import React, {Component} from 'react';
 import Footer from './Footer';
 import socket from '../socket';
-import MuiDialogTitle from '@material-ui/core/DialogTitle';
-import MuiDialogContent from '@material-ui/core/DialogContent';
-import MuiDialogActions from '@material-ui/core/DialogActions';
 import {
   Input,
   Button,
   Container,
   Dialog,
-  IconButton,
-  Typography,
   AppBar,
   Toolbar
 } from '@material-ui/core';
 
-import {withStyles} from '@material-ui/core/styles';
-import CloseIcon from '@material-ui/icons/Close';
-import {HowToPlay, About} from '.';
-
-const styles = theme => ({
-  root: {
-    margin: 0,
-    padding: theme.spacing(2)
-  },
-  closeButton: {
-    position: 'absolute',
-    right: theme.spacing(1),
-    top: theme.spacing(1),
-    color: theme.palette.grey[500]
-  }
-});
-
-const DialogTitle = withStyles(styles)(props => {
-  const {children, classes, onClose} = props;
-  return (
-    <MuiDialogTitle disableTypography className={classes.root}>
-      <Typography variant="h6">{children}</Typography>
-      {onClose ? (
-        <IconButton
-          aria-label="Close"
-          className={classes.closeButton}
-          onClick={onClose}
-        >
-          <CloseIcon />
-        </IconButton>
-      ) : null}
-    </MuiDialogTitle>
-  );
-});
-
-const DialogContent = withStyles(theme => ({
-  root: {
-    padding: theme.spacing(2)
-  }
-}))(MuiDialogContent);
-
-const DialogActions = withStyles(theme => ({
-  root: {
-    margin: 0,
-    padding: theme.spacing(1)
-  }
-}))(MuiDialogActions);
+import {HowToPlay, About, DialogActions} from '.';
 
 class Welcome extends Component {
   constructor() {
