+++ client/components/index.js
@@ -10,3 +10,4 @@ export {default as About} from './About';
 export {default as GameControlPanel} from './GameControlPanel';
 export {default as GameOver} from './GameOver';
 export {default as HowToPlay} from './HowToPlay';
+export {DialogTitle, DialogContent, DialogActions} from './Dialog';
